feat: sell to customer & collect money

+++ Assets/Juice Mama/Scripts/ScriptableObjects/PlayerData.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Data/PlayerSO")]
 public class PlayerData : ScriptableObject
 {
-    public int money;
+    public float money;
 }
